part04 till Ex21

+++ part4/controllers/blogs.js
@@ -1,9 +1,11 @@
+require('dotenv').config()
 const blogsRouter = require('express').Router()
-const { request } = require('../app')
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 blogsRouter.get('/', async (request, response) => {
-    const blogs = await Blog.find({})
+    const blogs = await Blog.find({}).populate('creator', {username: 1, name: 1, id: 1})
     response.json(blogs)
   })
 
@@ -24,16 +26,38 @@ blogsRouter.put('/:id', async (request, response) => {
   
 })
 blogsRouter.delete('/:id', async(request, response) => {
+  const decodedToken = jwt.verify(request.token, `${process.env.SECRET}`)
+  
+  const blogToDelete = await Blog.findById(request.params.id)
+  if (blogToDelete.creator.toString() === decodedToken.id.toString()) {
+    await Blog.findByIdAndRemove(request.params.id)
+    response.status(204).end()
+  } else {
+    response.status(401).json({error: "Only creator can delete"})
+  }
+  // await Blog.findByIdAndRemove(request.params.id)
 
-  await Blog.findByIdAndRemove(request.params.id)
-
-  response.status(204).end()
+  // response.status(204).end()
 
 })
 
+// const getTokenFrom = (request) => {
+//   const authorization = request.get('Authorization')
+//   if (authorization && authorization.startsWith('Bearer ')) {
+//     return authorization.replace('Bearer ', '')
+//   } else {
+//   return null
+//   }
+// }
 
 blogsRouter.post('/', async (request, response) => {
     const body = request.body
+    const decodedToken = jwt.verify(request.token, `${process.env.SECRET}`)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'invalid token'})
+    }
+
+    const user = (await User.findById(decodedToken.id))
 
     if (body.title === undefined) {
       return response.status(400).json({error: "title missing"})
@@ -47,10 +71,14 @@ blogsRouter.post('/', async (request, response) => {
       title: body.title,
       author: body.author,
       url: body.url,
-      likes: body.likes || 0
+      likes: body.likes || 0,
+      creator: user._id
     })
   
     const savedBlog = await blog.save()
+    user.blogs = user.blogs.concat(savedBlog.id)
+    
+    await user.save()
     response.status(201).json(savedBlog)
   })
